+++ flow.js
@@ -1,5 +1,3 @@
-/* eslint-disable import/no-commonjs */
-
 module.exports = {
   // warning, this is for tests
   // you must declare this in your config as it's inherited by "extends"