+++ index.js
@@ -1,6 +1,8 @@
-/* eslint-disable import/no-commonjs */
-
 module.exports = {
+  // warning, this is for tests
+  // you must declare this in your config as it's inherited by "extends"
+  parser: "babel-eslint",
+
   // mostly good stuff
   extends: "eslint:recommended",
 
@@ -51,13 +53,16 @@ module.exports = {
 
     "import/default": [ 2 ],
     "import/export": [ 2 ],
-    "import/extensions": [ 2 ],
+    "import/extensions": [ 0 ], // not flexible enough
     "import/imports-first": [ 2 ],
+    "import/max-dependencies": [ 2 ],
     "import/named": [ 2 ],
     "import/namespace": [ 2 ],
     "import/newline-after-import": [ 2 ],
+    "import/no-absolute-path": [ 2 ],
     "import/no-amd": [ 2 ],
-    "import/no-commonjs": [ 2 ],
+    // Intended for temporary use when migrating to pure ES6 modules.
+    "import/no-commonjs": [ 0 ],
     "import/no-deprecated": [ 2 ],
     "import/no-duplicates": [ 2 ],
     "import/no-extraneous-dependencies": [ 2 ],
@@ -66,9 +71,9 @@ module.exports = {
     "import/no-named-as-default": [ 2 ],
     "import/no-namespace": [ 2 ],
     "import/no-nodejs-modules": [ 0 ],
+    "import/no-restricted-paths": [ 2 ],
     "import/no-unresolved": [ 2, { "commonjs": true } ],
     "import/order": [ 2, { "newlines-between": "always" } ],
-    // not landed yet
-    // "import/prefer-default-export": [ 2 ],
+    "import/prefer-default-export": [ 0 ], // does not always make sense
   },
 }