+++ react-flow.js
@@ -1,7 +1,3 @@
-// for the url
-/* eslint-disable max-len */
-/* eslint-disable import/no-commonjs */
-
 module.exports = {
   // warning, this is for tests
   // you must declare this in your config as it's inherited by "extends"
@@ -15,21 +11,4 @@ module.exports = {
     "./react.js",
     "./flow.js",
   ],
-
-  rules: {
-    "react/sort-comp": [ 2, {
-      "order": [
-        // allow to add props/state on top of react classes
-        "props",
-        "state",
-
-        // defaults
-        // see https://github.com/yannickcr/eslint-plugin-react/blob/master/lib/rules/sort-comp.js
-        "static-methods",
-        "lifecycle",
-        "everything-else",
-        "render",
-      ],
-    } ],
-  },
 }